move e2e tests to different files
